% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin_table.R
\name{beginTable}
\alias{beginTable}
\title{Function to begin table within LaTeX document}
\usage{
beginTable(aligns)
}
\arguments{
\item{aligns}{This argument indicates the alignment to
be used in each column and the vertical lines to insert.
See https://en.wikibooks.org/wiki/LaTeX/Tables for more
information (character).}
}
\description{
This function begins a table within a LaTeX document.
}
