% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{formatNum}
\alias{formatNum}
\title{Function to do format numbers within LaTeX document.}
\usage{
formatNum(x, dec = 4, big.mark = ",", stat = F, se = F, pvec = -1)
}
\arguments{
\item{x}{The number(s) to be formatted (numeric).}

\item{dec}{Number of decimal places (numeric).}

\item{big.mark}{Used as mark within large numbers; normally,
a comma (as in 1,201,390) (character).}

\item{stat}{(logical)}

\item{se}{(logical)}

\item{pvec}{(numeric)}
}
\description{
This function formats numbers in a table within a
 LaTeX document.
}
