% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namerow.R
\name{namerow}
\alias{namerow}
\title{Function to label rows.}
\usage{
namerow(name, num)
}
\arguments{
\item{name}{The name of the rows (character).}

\item{num}{The number of &'s (numeric).}
}
\description{
This function creates LaTeX code to name rows.
}
