% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numrow.R
\name{numrow}
\alias{numrow}
\title{Function to do something in LaTeX document}
\usage{
numrow(title, array, dec = 2, stat = F, se = F, pvec = -1,
  spacer = "\\\\\\\\")
}
\arguments{
\item{title}{The name of the row (character).}

\item{array}{The array that will be formatted in the row (numeric).}

\item{dec}{The number of decimal places (numeric).}

\item{stat}{(logical)}

\item{se}{(logical)}

\item{pvec}{(numeric)}

\item{spacer}{LaTeX code for space between rows (character).}
}
\description{
This does something in a LaTeX document, but I
 am not exactly sure what, and I didn't want to
 intentionally write something untrue
}
