% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partmidrule.R
\name{partmidrule}
\alias{partmidrule}
\title{Function to draw horizontal lines  in LaTeX.}
\usage{
partmidrule(start, stop)
}
\arguments{
\item{start}{The column where the line begins (numeric).}

\item{stop}{The column where the line ends (numeric).}
}
\description{
This function draws horizontal lines across the columns specified.
}
