% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textables-tabular-rows.R
\name{row_header}
\alias{row_header}
\title{Function to create title row in LaTeX document.}
\usage{
row_header(panel, title, num)
}
\arguments{
\item{panel}{The name of the panel (character).}

\item{title}{The title of the row (character).}

\item{num}{(numeric).}
}
\description{
This function creates a title row in LaTeX.
}
