% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textables-tabular-rows.R
\name{row_numeric}
\alias{row_numeric}
\title{Function to make a numerical row in LaTeX document}
\usage{
row_numeric(title, array, dec = 2, percentage = F, se = F, pvalues = -1,
  spacer = "\\\\\\\\")
}
\arguments{
\item{title}{The name of the row (character).}

\item{array}{The array that will be formatted in the row (numeric).}

\item{dec}{The number of decimal places (numeric).}

\item{percentage}{(logical)}

\item{se}{(logical)}

\item{pvalues}{(numeric)}

\item{spacer}{LaTeX code for space between rows (character).}
}
\description{
This creates a numerical row, with options to treat the row
 as a percentage, an estimate with number of stars 
 determined by p-value (to indicate significance),
 or wrapped in parentheses (like a standard error).
}
