% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textables-tabular-formatting.R
\name{rule_mid_partial}
\alias{rule_mid_partial}
\title{Function to draw partial horizontal line in LaTeX.}
\usage{
rule_mid_partial(start, stop)
}
\arguments{
\item{start}{The column where the line begins (numeric).}

\item{stop}{The column where the line ends (numeric).}
}
\description{
This function draws horizontal lines across the columns specified.
}
