% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringrow.R
\name{stringrow}
\alias{stringrow}
\title{Function to create rows of text in LaTeX table.}
\usage{
stringrow(title, array, spacer = "\\\\\\\\")
}
\arguments{
\item{title}{The title of the row (character).}

\item{array}{The array that contains the information
that will be in the row.}

\item{spacer}{The LaTeX code for spacing between this
row and the next row (character).}
}
\description{
This function creates rows of text (in
 characters) in a LaTeX table.
}
