% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textables-tabular-formatting.R
\name{tabular_topbottom_pretty}
\alias{tabular_topbottom_pretty}
\title{Function to begin table within LaTeX document}
\usage{
tabular_topbottom_pretty(tabular, alignments)
}
\arguments{
\item{alignments}{This argument indicates the alignment to
be used in each column and the vertical lines to insert.
See https://en.wikibooks.org/wiki/LaTeX/Tables for more
information (character).}
}
\description{
This function begins a tabular within a LaTeX document.
}
