% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt-formatNum.R
\name{tt_formatNum}
\alias{tt_formatNum}
\title{Function to do format numbers within LaTeX document.}
\usage{
tt_formatNum(x, dec = 4, big.mark = ",", percentage = F, se = F,
  pvalues = NULL, surround = "")
}
\arguments{
\item{x}{The number(s) to be formatted (numeric).}

\item{dec}{Number of decimal places (numeric).}

\item{big.mark}{Used as mark within large numbers; normally,
a comma (as in 1,201,390) (character).}

\item{percentage}{(logical)}

\item{se}{(logical)}

\item{pvalues}{(numeric)}

\item{surround}{(string) allows to surround the results with text, ex "(%s)"}
}
\description{
This function formats numbers in a table within a
 LaTeX document.
}
