% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt-all.R
\name{tt_save}
\alias{tt_save}
\title{Save a textable object to .tex, with stand_alone option}
\usage{
tt_save(tabular, filename, stand_alone = F)
}
\description{
Save a textable object to .tex, with stand_alone option
}
