% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textables-write-tex.R
\name{write_tex}
\alias{write_tex}
\title{Function to end LaTeX document}
\usage{
write_tex(tabular, filename, stand_alone = F)
}
\arguments{
\item{tabular}{A tabular object in LaTeX.}

\item{filename}{Name of file (ending in .tex).}

\item{stand_alone}{Allows the tabular to be compiled directly by latex.}
}
\description{
This function ends a LaTeX document.
}
